requirements.txt
fastapi
uvicorn
pyjwt[crypto]
PyYAML
